+++ apps/io-func-wallet-solution-backend/src/app/config.ts
@@ -5,11 +5,17 @@ import * as RE from "fp-ts/lib/ReaderEither";
 import { sequenceS } from "fp-ts/lib/Apply";
 import { validate } from "../validation";
 import { FederationEntityMetadata } from "../entity-configuration";
-import {
-  CryptoConfiguration,
-  getCryptoConfigFromEnvironment,
-} from "../infra/crypto/config";
+
 import { readFromEnvironment } from "../infra/env";
+import { Jwk, fromBase64ToJwks } from "../jwk";
+
+export const CryptoConfiguration = t.type({
+  jwks: t.array(Jwk),
+  jwtDuration: t.string,
+  jwtDefaultAlg: t.string,
+});
+
+export type CryptoConfiguration = t.TypeOf<typeof CryptoConfiguration>;
 
 export const Config = t.type({
   federationEntity: FederationEntityMetadata,
@@ -51,4 +57,25 @@ const getFederationEntityConfigFromEnvironment: RE.ReaderEither<
       "Federation entity configuration is invalid"
     )
   )
+);
+
+export const getCryptoConfigFromEnvironment: RE.ReaderEither<
+  NodeJS.ProcessEnv,
+  Error,
+  CryptoConfiguration
+> = pipe(
+  sequenceS(RE.Apply)({
+    jwks: pipe(
+      readFromEnvironment("WalletKeys"),
+      RE.chainEitherKW(fromBase64ToJwks)
+    ),
+    jwtDuration: pipe(
+      readFromEnvironment("JwtDuration"),
+      RE.orElse(() => RE.right("1h"))
+    ),
+    jwtDefaultAlg: pipe(
+      readFromEnvironment("JwtDefaultAlg"),
+      RE.orElse(() => RE.right("ES256"))
+    ),
+  })
 );
